fix: get rid of withRouter()
