+++ src/views/NotImplemented.js
@@ -1,34 +1,32 @@
-import { Component } from 'react';
-import { withRouter } from 'react-router-dom';
+import { useLocation, useParams } from 'react-router-dom';
+
 import { AppLink } from '../components';
 
 /**
  * Boilerplate for non-implemented Views
  */
-export class NotImplementedView extends Component {
-  render() {
-    const { name, location, match } = this.props;
-    const componentName = name || this.displayName || this._reactInternalFiber.elementType.name || 'View';
-    const paramId = match.params?.id;
+const NotImplementedView = ({ name }) => {
+  const location = useLocation();
+  const { id: paramId } = useParams();
+  const componentName = name || 'View';
 
-    return (
-      <div>
-        <h1>{componentName} is under construction</h1>
-        <p>
-          This view is not implemented yet. Go to <AppLink to="/">home page</AppLink>
-        </p>
-        <p>
-          You've called the <b>{location?.pathname}</b> url
-          {paramId && (
-            <span>
-              {' '}
-              where <b>{paramId}</b> is a parameter
-            </span>
-          )}
-        </p>
-      </div>
-    );
-  }
-}
+  return (
+    <div>
+      <h1>{componentName} is under construction</h1>
+      <p>
+        This view is not implemented yet. Go to <AppLink to="/">home page</AppLink>
+      </p>
+      <p>
+        You've called the <b>{location?.pathname}</b> url
+        {paramId && (
+          <span>
+            {' '}
+            where <b>{paramId}</b> is a parameter
+          </span>
+        )}
+      </p>
+    </div>
+  );
+};
 
-export default withRouter(NotImplementedView);
+export default NotImplementedView;
